requirements.txt
fastapi
uvicorn
python-docx
openai
aiofiles
python-dotenv